requirements.txt
asgiref==3.7.2
Django==4.2.6
et-xmlfile==1.1.0
openpyxl==3.1.2
psycopg2-binary==2.9.9
python-decouple==3.8
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
